requirements.txt
asgiref==3.5.2
certifi @ file:///croot/certifi_1665076670883/work/certifi
Django==3.2.16
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.4.0
